backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
PyJWT
passlib[bcrypt]
psycopg2-binary
